feat: export initializeServer for testing
