+++ codecov.yml
@@ -2,6 +2,6 @@ coverage:
   status:
     project:
       default:
-        target: '80%'
+        target: '75%'
     patch:
       enabled: false
