+++ packages/bottender/src/cli/providers/messenger/__tests__/listPersona.spec.ts
@@ -1,13 +1,13 @@
 import { MessengerClient } from 'messaging-api-messenger';
 
-import getChannelConfig from '../../../shared/getChannelConfig';
+import getChannelConfig from '../../../../shared/getChannelConfig';
 import { listPersona } from '../persona';
-import * as log from '../../../shared/log';
+import * as log from '../../../../shared/log';
 
 jest.mock('messaging-api-messenger');
 
-jest.mock('../../../shared/log');
-jest.mock('../../../shared/getChannelConfig');
+jest.mock('../../../../shared/log');
+jest.mock('../../../../shared/getChannelConfig');
 
 const MOCK_FILE_WITH_PLATFORM = {
   channels: {
