+++ packages/bottender/src/cli/providers/telegram/__tests__/setWebhook.spec.ts
@@ -1,17 +1,17 @@
 import Confirm from 'prompt-confirm';
 import { TelegramClient } from 'messaging-api-telegram';
 
-import getChannelConfig from '../../../shared/getChannelConfig';
-import getWebhookFromNgrok from '../../../shared/getWebhookFromNgrok';
+import getChannelConfig from '../../../../shared/getChannelConfig';
+import getWebhookFromNgrok from '../../../../shared/getWebhookFromNgrok';
 import { setWebhook } from '../webhook';
-import * as log from '../../../shared/log';
+import * as log from '../../../../shared/log';
 
 jest.mock('messaging-api-telegram');
 jest.mock('prompt-confirm');
 
-jest.mock('../../../shared/getWebhookFromNgrok');
-jest.mock('../../../shared/log');
-jest.mock('../../../shared/getChannelConfig');
+jest.mock('../../../../shared/getWebhookFromNgrok');
+jest.mock('../../../../shared/log');
+jest.mock('../../../../shared/getChannelConfig');
 
 const MOCK_FILE_WITH_PLATFORM = {
   channels: {
