+++ packages/bottender/src/shared/__tests__/getChannelConfig.spec.ts
